requirements.txt
Django>=3.2,<4.0
python-dotenv==0.19.2
requests==2.26.0
PyJWT==1.7.1
